requirements.txt
Django==4.2.24
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-barcode==0.15.1
django-ckeditor==6.7.0
django-ckeditor-uploader==0.2.4
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0
openpyxl==3.1.2
qrcode==7.4.2
mysqlclient==2.2.0
django-jazzmin==2.6.0
django-extensions==3.2.3